requirements.txt
adbc_driver_manager==0.11.0
adbc_driver_postgresql==0.11.0
adbc_driver_sqlite==0.11.0
AppKit==0.2.8
attr==0.3.2
beautifulsoup4==4.12.3
blinker==1.6.2
boto3==1.34.88
botocore==1.34.88
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
dask==2024.4.2
dl==0.1.0
docutils==0.21.1
fastparquet==2023.10.1
filelock==3.13.4
Foundation==0.1.0a0.dev1
fsspec==2023.10.0
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.100.1
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
js==1.0
keyring==24.3.1
MarkupSafe==2.1.2
matplotlib==3.8.4
mock==5.1.0
mtrand==0.1
mypy==1.9.0
numarray==1.5.1
Numeric==24.2
numexpr==2.10.0
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.2
pickle5==0.0.12
protobuf==5.26.1
psutil==5.9.8
py==1.11.0
PyInstaller==6.6.0
PyJWT==2.8.0
pyodide==0.0.2
pyOpenSSL==24.1.0
PyQt4==4.11.4
PyQt5==5.15.10
PyQt6==6.6.1
PySide6==6.7.0
PySimpleGUI==5.0.4
PySimpleGUI==5.0.4
pytest==8.1.1
python_calamine==0.2.0
python_magic==0.4.27
pyxlsb==1.0.10
PyYAML==6.0.1
qtpy==2.4.1
railroad==0.5.0
redis==5.0.3
requests_futures==1.0.1
requests_jwt==0.6.0
requests_kerberos==0.14.0
s3fs==2024.3.1
scikit_learn==1.4.2
scipy==1.13.0
sets==0.3.2
simplejson==3.19.2
Sphinx==7.3.7
SQLAlchemy==2.0.29
tables==3.9.2
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.3
trove_classifiers==2024.4.10
urllib3_secure_extra==0.1.0
xarray==2024.3.0
xlrd==2.0.1
xlsxwriter==3.2.0
xmlrpclib==1.0.1
zstandard==0.22.0

pandas~=2.1.2
python-dateutil~=2.8.2
jira~=3.6.0